Use context to reduce request stub setup
